220 Eastern Parkway; Louisville, KY 40209
The J. B. Speed School of Engineering was founded
in 1925 as part of the University of Louisville in
the U.S. state of Kentucky with money from the James
Breckenridge Speed Foundation which was created by
his children William S. Speed and Olive Speed Sackett
to honor J. B. Speed. Until 2003, it was known as the
J. B. Speed Scientific School.

The school is colloquially referred to as Speed School
or just "Speed" by students and locals.

https://en.wikipedia.org/wiki/J._B._Speed_School_of_Engineering